[bug-40] Group buttons under a folder
