--- src/core/hero/google-download-button.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { Box, Grid } from '@material-ui/core'
-import GoogleStoreDownload from 'core/hero/google-store.png'
-
-export const GoogleDownloadButton = () => (
-  <Grid item>
-    <Box m={1.25}>
-      <a href={'#'}>
-        <img src={GoogleStoreDownload} alt={'Descarcă aplicația de pe Google Play'} />
-      </a>
-    </Box>
-  </Grid>
-)
